Refactor: factor out tests.rs from push_pull
